refactor: extract story management functions to lib/story-ops.ts
Extracted getStories, getCurrentStory, formatStoryForTemplate,
formatCompletedStories, and parseAndInsertStories to a dedicated
module. This reduces step-ops.ts complexity and improves code
organization.

step-ops.ts: 1202 -> ~920 lines
Backwards compatibility maintained via re-export
