+++ src/installer/step-ops.ts
@@ -1,5 +1,6 @@
 import { getDb } from "../db.js";
 import type { LoopConfig, Story } from "./types.js";
+import { getStories, getCurrentStory, formatStoryForTemplate, formatCompletedStories, parseAndInsertStories } from "../lib/story-ops.js";
 import fs from "node:fs";
 import path from "node:path";
 import os from "node:os";
@@ -149,117 +150,7 @@ function readProgressFile(runId: string): string {
 /**
  * Get all stories for a run, ordered by story_index.
  */
-export function getStories(runId: string): Story[] {
-  const db = getDb();
-  const rows = db.prepare(
-    "SELECT * FROM stories WHERE run_id = ? ORDER BY story_index ASC"
-  ).all(runId) as any[];
-  return rows.map(r => ({
-    id: r.id,
-    runId: r.run_id,
-    storyIndex: r.story_index,
-    storyId: r.story_id,
-    title: r.title,
-    description: r.description,
-    acceptanceCriteria: JSON.parse(r.acceptance_criteria),
-    status: r.status,
-    output: r.output ?? undefined,
-    retryCount: r.retry_count,
-    maxRetries: r.max_retries,
-  }));
-}
-
-/**
- * Get the story currently being worked on by a loop step.
- */
-export function getCurrentStory(stepId: string): Story | null {
-  const db = getDb();
-  const step = db.prepare(
-    "SELECT current_story_id FROM steps WHERE id = ?"
-  ).get(stepId) as { current_story_id: string | null } | undefined;
-  if (!step?.current_story_id) return null;
-  const row = db.prepare("SELECT * FROM stories WHERE id = ?").get(step.current_story_id) as any;
-  if (!row) return null;
-  return {
-    id: row.id,
-    runId: row.run_id,
-    storyIndex: row.story_index,
-    storyId: row.story_id,
-    title: row.title,
-    description: row.description,
-    acceptanceCriteria: JSON.parse(row.acceptance_criteria),
-    status: row.status,
-    output: row.output ?? undefined,
-    retryCount: row.retry_count,
-    maxRetries: row.max_retries,
-  };
-}
-
-function formatStoryForTemplate(story: Story): string {
-  const ac = story.acceptanceCriteria.map((c, i) => `  ${i + 1}. ${c}`).join("\n");
-  return `Story ${story.storyId}: ${story.title}\n\n${story.description}\n\nAcceptance Criteria:\n${ac}`;
-}
-
-function formatCompletedStories(stories: Story[]): string {
-  const done = stories.filter(s => s.status === "done");
-  if (done.length === 0) return "(none yet)";
-  return done.map(s => `- ${s.storyId}: ${s.title}`).join("\n");
-}
-
-// ── T5: STORIES_JSON parsing ────────────────────────────────────────
-
-/**
- * Parse STORIES_JSON from step output and insert stories into the DB.
- */
-function parseAndInsertStories(output: string, runId: string): void {
-  const lines = output.split("\n");
-  const startIdx = lines.findIndex(l => l.startsWith("STORIES_JSON:"));
-  if (startIdx === -1) return;
-
-  // Collect JSON text: first line after prefix, then subsequent lines until next KEY: or end
-  const firstLine = lines[startIdx].slice("STORIES_JSON:".length).trim();
-  const jsonLines = [firstLine];
-  for (let i = startIdx + 1; i < lines.length; i++) {
-    if (/^[A-Z_]+:\s/.test(lines[i])) break;
-    jsonLines.push(lines[i]);
-  }
-
-  const jsonText = jsonLines.join("\n").trim();
-  let stories: any[];
-  try {
-    stories = JSON.parse(jsonText);
-  } catch (e) {
-    throw new Error(`Failed to parse STORIES_JSON: ${(e as Error).message}`);
-  }
-
-  if (!Array.isArray(stories)) {
-    throw new Error("STORIES_JSON must be an array");
-  }
-  if (stories.length > 20) {
-    throw new Error(`STORIES_JSON has ${stories.length} stories, max is 20`);
-  }
-
-  const db = getDb();
-  const now = new Date().toISOString();
-  const insert = db.prepare(
-    "INSERT INTO stories (id, run_id, story_index, story_id, title, description, acceptance_criteria, status, retry_count, max_retries, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, 'pending', 0, 2, ?, ?)"
-  );
-
-  const seenIds = new Set<string>();
-  for (let i = 0; i < stories.length; i++) {
-    const s = stories[i];
-    // Accept both camelCase and snake_case
-    const ac = s.acceptanceCriteria ?? s.acceptance_criteria;
-    if (!s.id || !s.title || !s.description || !Array.isArray(ac) || ac.length === 0) {
-      throw new Error(`STORIES_JSON story at index ${i} missing required fields (id, title, description, acceptanceCriteria)`);
-    }
-    if (seenIds.has(s.id)) {
-      throw new Error(`STORIES_JSON has duplicate story id "${s.id}"`);
-    }
-    seenIds.add(s.id);
-    insert.run(crypto.randomUUID(), runId, i, s.id, s.title, s.description, JSON.stringify(ac), now, now);
-  }
-}
+// Story functions now imported from ../lib/story-ops.ts
 
 // ── Abandoned Step Cleanup ──────────────────────────────────────────
 
@@ -1121,3 +1012,6 @@ export async function failStep(stepId: string, error: string): Promise<{ retryin
     return { retrying: true, runFailed: false };
   }
 }
+
+// Re-export story functions for backwards compatibility
+export { getStories, getCurrentStory, formatStoryForTemplate, formatCompletedStories, parseAndInsertStories } from "../lib/story-ops.js";
